requirements.txt
captum==0.6.0
matplotlib==3.6.2
numpy==1.23.3
opencv_python_headless==4.7.0.68
pandas==1.5.2
Pillow==9.4.0
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.64.1
